Creazione Manuali/requirements.txt
google-api-python-client
google-auth
google-auth-httplib2
openai
PyPDF2
Flask 